Update titles, error handling
Set the title of a detailed post page to "Parellagram - Title" instead
of just Parellagram, and crash on inability to load styles, instead of
just passing back error. In that case, something's wrong and should be
fixed, not ignored.

Also redirect when someone tries to access a page that is not
successfully parsed, although in general unless they're trying to do it
on purpose that shouldn't be possible.