Write artifacts to system temp directory
Instead of having an artifacts directory cluttering the project like I
had previously, write artifacts to the tempdir of the operating system
that the program is being run on. Blows away that directory on start
up, so hopefully nobody decides to put anything in there that they
shouldn't. Which would be anything. Don't put anything of your own
in there.