Make port configurable, add file watching
Program now watches the directory specified in config for changes, and
rebuilds the website on each change. Could be modified to only update
the landing page and the specific page that is changed, should change
that in the future.

Also made the port that is listened on a configurable value. Much easier
to put this behind a server and not have to run it with super user
privilges because lord knows what might happen then.